pyproject.toml
[tool.poetry]
name = "channelsorter"
version = "0.1.0"
description = ""
authors = ["Slavfox <slavfoxman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
tortoise-orm = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
